requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.2
Django==4.1.6
djangorestframework==3.14.0
h11==0.13.0
idna==3.3
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
outcome==1.1.0
pycparser==2.21
pygame==2.1.2
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.7.1
rethinkdb==2.4.8
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.3
thinker==1.1.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.9
wsproto==1.1.0
